+++ source/Energinet.DataHub.MarketParticipant.IntegrationTests/Hosts/WebApi/UserRoleAuditLogIntegrationTests.cs
@@ -63,7 +63,7 @@ public sealed class UserRoleAuditLogIntegrationTest : WebApiIntegrationTestsBase
             name,
             "description",
             UserRoleStatus.Active,
-            EicFunction.Agent,
+            EicFunction.BillingAgent,
             new Collection<int>() { (int)Permission.ActorManage });
 
         var createUserRoleCommand = new CreateUserRoleCommand(userId, createUserRoleDto);
@@ -97,7 +97,7 @@ public sealed class UserRoleAuditLogIntegrationTest : WebApiIntegrationTestsBase
 
         var userRoleId = await _fixture
             .DatabaseManager
-            .CreateUserRoleAsync("UpdateUserRole_NameChangedAuditLog_InitName", string.Empty, UserRoleStatus.Active, EicFunction.Agent, new[] { Permission.UsersView });
+            .CreateUserRoleAsync("UpdateUserRole_NameChangedAuditLog_InitName", string.Empty, UserRoleStatus.Active, EicFunction.BillingAgent, new[] { Permission.UsersView });
 
         const string nameUpdate = "UpdateUserRole_NameChangedAuditLog_Updated";
         const string descriptionUpdate = "UpdateUserRole_DescriptionChangedAuditLog_Updated";
@@ -218,7 +218,7 @@ public sealed class UserRoleAuditLogIntegrationTest : WebApiIntegrationTestsBase
             name,
             "description",
             UserRoleStatus.Active,
-            EicFunction.Agent,
+            EicFunction.BillingAgent,
             new Collection<int>() { (int)Permission.ActorManage });
     }
 
@@ -248,7 +248,7 @@ public sealed class UserRoleAuditLogIntegrationTest : WebApiIntegrationTestsBase
             updateUserRoleDto.Description,
             updateUserRoleDto.Status,
             updateUserRoleDto.Permissions.Select(x => (Permission)x),
-            EicFunction.Agent);
+            EicFunction.BillingAgent);
 
         return userRoleAuditLogService.BuildAuditLogsForUserRoleCreated(
             new UserId(userId ?? Guid.NewGuid()),