+++ source/Energinet.DataHub.MarketParticipant.Tests/Services/BusinessRoleCodeDomainServiceTests.cs
@@ -78,7 +78,7 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Services
 
             var markedRoles = new[]
             {
-                EicFunction.EnergyTrader,
+                EicFunction.BillingAgent,
                 EicFunction.EnergySupplier
             };